Store board climbs on DB instead of having to fetch everytime.
Now board climbs are indexed to see if they have already been logged on the app. If they have than it uses the key value pair to get the name. Else it fetches the value from the API and stores it.

TODO: Merge current sessions with synced sessions from tension.